// generated/backend/models/Note.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { NoteTopology } from './NoteTopology';
export type Note = {
    noteTopology: NoteTopology;
    details?: string;
    parentId?: number;
    updatedAt: string;
    id: number;
    createdAt: string;
    readonly deletedAt?: string;
    wikidataId?: string;
};